Fixed VS project file. Removed stuff from platform/msvc.
git-svn-id: https://love.svn.sf.net/svnroot/love/trunk@406 3494dbca-881a-0410-bd34-8ecbaf855390

+++ src/Font.cpp
@@ -35,6 +35,6 @@ namespace love
 	
 	float Font::getWidth(const char character) const
 	{
-		return widths[(int)character];
+		return (float)widths[(int)character];
 	}
 }
